% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_key_cols_unique.R
\name{check_key_cols_unique}
\alias{check_key_cols_unique}
\title{Check key cols unique}
\usage{
check_key_cols_unique(df, key_cols, verbose = FALSE)
}
\arguments{
\item{df}{a dataframe}

\item{key_cols}{vector of column names}

\item{verbose}{TRUE/FALSE should we print a message?}
}
\value{
TRUE if key cols have unique rows; FALSE if not
}
\description{
Check key cols unique
}
\examples{
irisint = iris
irisint$rownum = 1:nrow(irisint)
key_cols = c("rownum")
check_key_cols_unique(irisint, key_cols, TRUE)
check_key_cols_unique(irisint, "Species", TRUE)
}
