% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vec_class.R
\name{check_vec_class}
\alias{check_vec_class}
\title{Title}
\usage{
check_vec_class(x, required_class)
}
\arguments{
\item{x}{a vector}

\item{required_class}{a function checking the class, eg. is.factor}
}
\description{
Title
}
\examples{
\dontrun{
check_vec_class(iris$Species, is.factor)
}
}
