% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_df_diff_tables.R
\name{compute_df_diff_tables}
\alias{compute_df_diff_tables}
\title{Compute df diff tables}
\usage{
compute_df_diff_tables(new_df, old_df = NA, key_cols = NA, verbose = FALSE)
}
\arguments{
\item{new_df}{dataframe. A dataframe of new data}

\item{old_df}{dataframe. A dataframe of old data. new_df and old_df can have overlapping data}

\item{key_cols}{optional vector of column names that constitite a unique table key.}

\item{verbose}{boolean, default FALSE. Should the processing be chatty?}
}
\value{
a list of dataframes with the structure
list(
"new_rows" = new_rows,
"modified_rows" = modified_rows,
"deleted_rows" = deleted_rows
)
}
\description{
Returns a list of adds, updates and deletes required to transform old_df into new_df.
Part of the \code{dataset} family of functons in minio utils. These include \code{compute_df_diff_tables}, \code{update_minio_dataset}, \code{retrieve_dataset}, \code{retrieve_dataset_diffs} and \code{retrieve_dataset_diff_stats}.
}
\examples{
iris$key <- 1:nrow(iris)

old_df <- iris[1:100,]
old_df[75,1] <- 100
new_df <- iris[50:150,]
compute_df_diff_tables(new_df, old_df, key_cols = "key")
}
