% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_backup.R
\name{delete_backup}
\alias{delete_backup}
\title{Delete backup}
\usage{
delete_backup(destination, after_position, verbose = FALSE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{after_position}{how many backups should we leave in the dv?}

\item{verbose}{TRUE /FALSE should the function be chatty?}
}
\value{
TRUE
}
\description{
Delete backup
}
