% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroy_dv.R
\name{destroy_dv}
\alias{destroy_dv}
\title{Destroy dv}
\usage{
destroy_dv(destination, prompt = TRUE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{prompt}{should we ask for manual confirmation?}
}
\value{
TRUE
}
\description{
Destroy dv
}
\examples{
temp_dir <- tempfile()
dir.create(temp_dir, recursive = TRUE)
df <- data.frame(a = 1:5, b = letters[1:5])

create_dv(df, temp_dir)
destroy_dv(temp_dir, prompt = FALSE)

unlink(temp_dir)

}
