% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_generator.R
\name{documentation_generator}
\alias{documentation_generator}
\title{Documentation generator}
\usage{
documentation_generator(df, path)
}
\arguments{
\item{df}{a dataframe to create documentation for}

\item{path}{string. a file path to save the documentation json file at.}
}
\value{
TRUE
}
\description{
Documentation generator
}
\examples{
\dontrun{
documentation_generator(mtcars, "mtdocumentation.json")
}
}
