% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_path.R
\name{fix_path}
\alias{fix_path}
\title{Fix path}
\usage{
fix_path(path, destination, verbose = FALSE)
}
\arguments{
\item{path}{a sub prefix of the destination}

\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{verbose}{TRUE /FALSE should the function be chatty?}
}
\value{
an arrow SubTreeFileSystem
}
\description{
Fix path
}
