% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_backups.R
\name{get_backups}
\alias{get_backups}
\title{Get backups}
\usage{
get_backups(destination, collect = TRUE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{collect}{should we collect the underlying arrow dataset or return just the connection?}
}
\value{
an arrow dataset
}
\description{
Get backups
}
