% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diffs.R
\name{get_diffs}
\alias{get_diffs}
\title{Get diffs}
\usage{
get_diffs(destination, collect = TRUE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{collect}{should we collect the underlying arrow dataset or return just the connection?}
}
\value{
a data frame or an arrow dataset connection
}
\description{
Get diffs
}
