% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest.R
\name{get_latest}
\alias{get_latest}
\title{Get latest}
\usage{
get_latest(destination, collect = TRUE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{collect}{should we return a data frame (TRUE) or an arrow dataset connection (FALSE)}
}
\value{
a data frame or an arrow dataset
}
\description{
Get latest
}
