% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(destination, verbose = FALSE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{verbose}{TRUE /FALSE should the function be chatty?}
}
\value{
a list
}
\description{
Get metadata
}
