% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash_precursor}
\alias{hash_precursor}
\title{A semi-vectorized hash function}
\usage{
hash_precursor(...)
}
\arguments{
\item{...}{values}
}
\value{
a vector of hashed values (if ... all have the same length), a scalar otherwise
}
\description{
If the ... arguments all have the same length, this function will return
a vector of the same length with hashes corresponding to each entry in the
corresponding list. If the ... arguments don't have the same length,
the overall hash of all values is returned.
}
