% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash_scalar}
\alias{hash_scalar}
\title{Hash one or more values (non-vectorized)}
\usage{
hash_scalar(...)
}
\arguments{
\item{...}{values}
}
\value{
a hash of the provided values
}
\description{
Hash one or more values (non-vectorized)
}
