% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prefix.R
\name{make_prefix}
\alias{make_prefix}
\title{Make prefix}
\usage{
make_prefix(destination)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
silent
}
\description{
Create the necessary subdirectories to make sure that SubTreeFileSystem methods will work.
}
\examples{
temp_dir <- tempfile()
make_prefix(temp_dir)

}
