% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SubTreeFileSystem.R
\name{make_SubTreeFileSystem}
\alias{make_SubTreeFileSystem}
\title{Make SubTreeFileSystem}
\usage{
make_SubTreeFileSystem(destination, verbose = FALSE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{verbose}{TRUE /FALSE should the function be chatty?}
}
\value{
an arrow SubTreeFileSystem
}
\description{
Make SubTreeFileSystem
}
