% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{precursor_hash_write}
\alias{precursor_hash_write}
\title{Write hashed value to file}
\usage{
precursor_hash_write(data_file_path, ..., precursor_hash = NULL)
}
\arguments{
\item{data_file_path}{path where the data is stored. ".hash" is appended
to the file name, and if that file exists, the lines are read and
compared to the value of the hashed object}

\item{...}{values to hash - unnecessary if precursor_hash is not null}

\item{precursor_hash}{hash of values to compare to hash saved in data_file_path.hash}
}
\description{
Assumes that the path provided is the path where the data created using the
precursor value is stored.
}
