% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_backup.R
\name{put_backup}
\alias{put_backup}
\title{Put backup}
\usage{
put_backup(new_df, destination)
}
\arguments{
\item{new_df}{a data frame}

\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
TRUE
}
\description{
Put backup
}
