% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_diff_stats.R
\name{put_diff_stats}
\alias{put_diff_stats}
\title{Put diff stats}
\usage{
put_diff_stats(destination)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
TRUE
}
\description{
Put diff stats
}
