% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_latest.R
\name{put_latest}
\alias{put_latest}
\title{Put latest}
\usage{
put_latest(new_df, destination)
}
\arguments{
\item{new_df}{a dataframe}

\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
TRUE
}
\description{
Put latest
}
