% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv_diff.R
\name{read_dv_diff}
\alias{read_dv_diff}
\title{Read dv diff}
\usage{
read_dv_diff(destination, as_of, key_cols)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{as_of}{the valid date at which you'd like to read the dv}

\item{key_cols}{a character vector of column names that constitute a unique key}
}
\value{
a data frame
}
\description{
Read dv diff
}
