% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_prefix.R
\name{remove_prefix}
\alias{remove_prefix}
\title{Remove prefix}
\usage{
remove_prefix(destination, prompt = TRUE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{prompt}{should we ask for user confirmation?}
}
\value{
TRUE
}
\description{
Remove prefix
}
