% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_dataset_documentation.R
\name{retrieve_dataset_documentation}
\alias{retrieve_dataset_documentation}
\title{Retrieve dataset documentation}
\usage{
retrieve_dataset_documentation(
  prefix,
  save_path = "documentation.json",
  open = FALSE,
  minio_key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  minio_secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  minio_url = "lake.capetown.gov.za",
  verbose = FALSE
)
}
\arguments{
\item{prefix}{string. s3 location of dataset, including bucket name}

\item{save_path}{string. Defaults to documentation.json. local path to save documentation file to.}

\item{open}{logocal. Defaults to FALSE. Should we open the file once we've retrieved it?}

\item{minio_key}{string. S3 access key. Defaults to Sys.getenv("AWS_ACCESS_KEY_ID").}

\item{minio_secret}{string. S3 secret key. Defaults to Sys.getenv("AWS_SECRET_ACCESS_KEY")}

\item{minio_url}{string. S3 url. Defaults to "lake.capetown.gov.za"}

\item{verbose}{TRUE/FALSE. Should the operations be chatty.}
}
\value{
TRUE
}
\description{
Retrieve dataset documentation
}
\examples{
\dontrun{
#'# Set AWS secrets
Sys.setenv(
  "AWS_ACCESS_KEY_ID" = MINIO_KEY,
 "AWS_SECRET_ACCESS_KEY" = MINIO_SECRET
)

retrieve_dataset_documentation("test_dataset")
                 }
}
