% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_dataset_stats.R
\name{retrieve_dataset_stats}
\alias{retrieve_dataset_stats}
\title{Retrieve dataset statistics (if the exists)}
\usage{
retrieve_dataset_stats(
  prefix,
  timestamp_filter = NA,
  collect = TRUE,
  minio_key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  minio_secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  minio_url = "lake.capetown.gov.za",
  verbose = FALSE
)
}
\arguments{
\item{prefix}{string. The dataset location.}

\item{timestamp_filter}{datetime. The timestamp of the time where you want to seent he summary stats for. If NA (default), returns the latest summary stats.}

\item{collect}{TRUE/FALSE. Do you want an arrow dataset connection, or a collected dataframe?}

\item{minio_key}{string. S3 access key. Defaults to Sys.getenv("AWS_ACCESS_KEY_ID").}

\item{minio_secret}{string. S3 secret key. Defaults to Sys.getenv("AWS_SECRET_ACCESS_KEY")}

\item{minio_url}{string. S3 url. Defaults to "lake.capetown.gov.za"}

\item{verbose}{TRUE/FALSE. Should the operations be chatty.}
}
\value{
a dataframe or arrow dataset connection
}
\description{
Retrieve dataset statistics (if the exists)
}
\examples{
\dontrun{
#'# Set AWS secrets
Sys.setenv(
  "AWS_ACCESS_KEY_ID" = MINIO_KEY,
 "AWS_SECRET_ACCESS_KEY" = MINIO_SECRET
)

retrieve_dataset_stats("test_dataset",
                 lubridate::now() - lubridate::days(90))
                 }
}
