% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{sorted_count}
\alias{sorted_count}
\title{Sorted count}
\usage{
sorted_count(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a string containing the counts of the 3 most numerous factor levels
}
\description{
Sorted count
}
\examples{
sorted_count(iris$Species)
}
