% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_df.R
\name{summarise_df}
\alias{summarise_df}
\title{Summarise a dataframe}
\usage{
summarise_df(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a dataframe with salient summarty statistics
}
\description{
Summarise a dataframe
}
\examples{
summarise_df(iris)
}
