% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_df_by_type.R
\name{summarise_df_by_type}
\alias{summarise_df_by_type}
\title{Title}
\usage{
summarise_df_by_type(df, cols, sum_funs)
}
\arguments{
\item{df}{a dataframe}

\item{cols}{string vector of column names}

\item{sum_funs}{a named vector of summary functions to apply to the columns}
}
\value{
a dataframe of summary statistics
}
\description{
Title
}
