% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_diff.R
\name{summarise_diff}
\alias{summarise_diff}
\title{Summarise diff}
\usage{
summarise_diff(destination)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
a data frame of diff statistics
}
\description{
Summarise diff
}
