% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_diffs.R
\name{summarise_diffs}
\alias{summarise_diffs}
\title{Summarise diffs}
\usage{
summarise_diffs(destination)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
a data frame of diff statistics
}
\description{
Summarise diffs
}
