% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{top_counts}
\alias{top_counts}
\title{Title}
\usage{
top_counts(x, max_char = 4, max_levels = 3)
}
\arguments{
\item{x}{a factor vector}

\item{max_char}{integer. number of characters to truncate the factors to}

\item{max_levels}{integer. the max number of levels to display}
}
\value{
a string with the shortened names of the top factors and the count of their occurrences.
}
\description{
Title
}
\examples{
top_counts(iris$Species, 4, 3)
}
