% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_summary_df.R
\name{transpose_summary_df}
\alias{transpose_summary_df}
\title{Transpose summary dataframe}
\usage{
transpose_summary_df(df)
}
\arguments{
\item{df}{a summary statistics dataframe}
}
\value{
a neatly transposed dataframe
}
\description{
Transpose summary dataframe
}
