% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dataset_documentation.R
\name{update_dataset_documentation}
\alias{update_dataset_documentation}
\title{Title}
\usage{
update_dataset_documentation(
  documentation_file = "documentation.json",
  prefix,
  minio_key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  minio_secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  minio_url = "lake.capetown.gov.za",
  verbose = FALSE
)
}
\arguments{
\item{documentation_file}{string. the place you want to save the documentation file.}

\item{prefix}{string. the bucket plus prefix of the location of the dataset.}

\item{minio_key}{string. S3 access key. Defaults to Sys.getenv("AWS_ACCESS_KEY_ID").}

\item{minio_secret}{string. S3 secret key. Defaults to Sys.getenv("AWS_SECRET_ACCESS_KEY")}

\item{minio_url}{string. S3 url. Defaults to "lake.capetown.gov.za"}

\item{verbose}{TRUE/FALSE. Should the operations be chatty.}
}
\value{
TRUE
}
\description{
Title
}
\examples{
\dontrun{
# Set AWS secrets
Sys.setenv(
  "AWS_ACCESS_KEY_ID" = MINIO_KEY,
 "AWS_SECRET_ACCESS_KEY" = MINIO_SECRET
)

update_dataset_documentation(prefix)
                    }
}
