% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dv.R
\name{update_dv}
\alias{update_dv}
\title{Update dv}
\usage{
update_dv(df, destination)
}
\arguments{
\item{df}{a data frame. Must be structurally identical to the dv.}

\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
TRUE
}
\description{
Update dv
}
