% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_diff.R
\name{commit_diff}
\alias{commit_diff}
\title{Commit diff}
\usage{
commit_diff(diff_df, destination, verbose = FALSE)
}
\arguments{
\item{diff_df}{a data frame. Output off diffdfs::diffdfs.}

\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{verbose}{TRUE /FALSE should the function be chatty?}
}
\value{
TRUE
}
\description{
Commit diff
}
