% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroy_dv.R
\name{destroy_dv}
\alias{destroy_dv}
\title{Destroy dv}
\usage{
destroy_dv(destination, prompt = TRUE)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{prompt}{should we ask for manual confirmation?}
}
\value{
TRUE
}
\description{
Destroy dv
}
