% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff_stats.R
\name{get_diff_stats}
\alias{get_diff_stats}
\title{Get diff stats}
\usage{
get_diff_stats(destination)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
a data frame
}
\description{
Get diff stats
}
