% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv_backup.R
\name{read_dv_backup}
\alias{read_dv_backup}
\title{Read dv backup}
\usage{
read_dv_backup(destination, as_of)
}
\arguments{
\item{destination}{a local directory path or an arrow SubTreeFileSystem}

\item{as_of}{the valid date at which you'd like to read the dv}
}
\description{
Read dv backup
}
