% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dv.R
\name{update_dv}
\alias{update_dv}
\title{Update dv}
\usage{
update_dv(df, destination)
}
\arguments{
\item{df}{a data frame. Must be structurally identical to the dv.}

\item{destination}{a local directory path or an arrow SubTreeFileSystem}
}
\value{
TRUE
}
\description{
Update a versioned dataset with a new version of the data.
}
\examples{
temp_dir <- tempfile()
dir.create(temp_dir, recursive = TRUE)
df <- data.frame(a = 1:5, b = letters[1:5])
new_df <- data.frame(a = 2:5, b = letters[2:5])
create_dv(df, temp_dir)

update_dv(new_df, temp_dir)

}
